Some code prepping and fixed some double integer in code

+++ src/game/WorldHandlers/GuildHandler.cpp
@@ -329,9 +329,10 @@ void WorldSession::HandleGuildInfoOpcode(WorldPacket& /*recvPacket*/)
 
     WorldPacket data(SMSG_GUILD_INFO, (guild->GetName().size() + 4 + 4 + 4));
     data << guild->GetName();
-    data << uint32(secsToTimeBitFields(guild->GetCreatedDate())); // 3.x (prev. day + month + year)
-    data << uint32(guild->GetMemberSize());                 // amount of chars
-    data << uint32(guild->GetAccountsNumber());             // amount of accounts
+    data << secsToTimeBitFields(guild->GetCreatedDate()); // 3.x (prev. day + month + year)
+    data << guild->GetMemberSize();                       // amount of chars
+    data << guild->GetAccountsNumber();                   // amount of accounts
+
     SendPacket(&data);
 }
